Surface rendering

+++ formex.py
@@ -35,7 +35,7 @@ def deprecated(old,new):
     print "Function %s is deprecated: use %s instead" % (old,new)
 
 # default float and int types used in the Formex data
-Float = float64
+Float = float32
 Int = int32
